Ressource adding, implementation and modification Part 12
Implemented the Wooden Pickaxe, added it as stone producing dependency (Ratio Stone:Pickaxe 7:1). and modified the stone [amountPerLevel] to 7 to fit that new ratio. I promise I'll start making the game this year... Maybe the "game" won't be ready until this year's December month... But I swear he'll be ready before the end of the year !